Add an 'unwrapped' navigation coordinator
Adds the 'UnwrappedNavigationViewCoordinator' as a 'convenience' class.
This class is a simplified version of the 'NavigationViewCoordinator',
with the only difference that it doesn't wrap the view in a
NavigationView.

Using this class over the 'NavigationViewCoordinator' results in a
'single' view on macOS, instead of macOS choosing to make it split
screen (due to the NavigationViewCoordinator's NavigationView).